Update README.md
Added new environment variable REQUIRE_API_KEY as well as the new /v1/voices/all endpoint and the fixed up /v1/voices endpoint. 

The Docker Hub version of this project has not been updated to account for the changes in the commit prior to this one with the changes described above. 

Will wait a bit and do some manual testing before pushing the update, despite very minimal risk.